Refactor pin code fields
