+++ lib/src/constants/enum/animations.dart
@@ -1,0 +1,11 @@
+enum Animations {
+  fade,
+  grow,
+  shrink,
+  slideInUp,
+  slideInDown,
+  slideInLeft,
+  slideInRight,
+  rotateRight,
+  rotateLeft,
+}
