+++ lib/src/constants/enum/field_border_style.dart
@@ -1,0 +1,9 @@
+enum FieldBorderStyle {
+  top,
+  bottom,
+  left,
+  right,
+  topBottom,
+  leftRight,
+  square,
+}
